requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0.0

# PyTorch with CUDA support for RTX 5070
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Optional: TensorFlow (if needed for additional models)
# tensorflow==2.13.0

# Machine learning utilities
scikit-learn==1.3.0

# Video processing
ffmpeg-python==0.2.0

# Hugging Face datasets for UCF101
datasets>=2.14.0
huggingface-hub>=0.16.0

# Windows-specific optimizations
psutil>=5.9.0  # System monitoring
GPUtil>=1.4.0  # GPU monitoring

# Development and debugging
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: Weights & Biases logging

# Built-in modules (no installation needed)
# argparse, pathlib, logging, json, os, sys, time, datetime
